Added one solution step.
A lot of the feedback on this page, and that I've seen elsewhere such as Reddit, come from people who get an error when trying to update HAOS. HA gives an unhealthy error if Home Assistant Supervisor needs to be updated, which has to happen before HAOS can be updated. This should provide direction for folks that are asking about how to start the D-Bus daemon.